+++ test/test-utils.ts
@@ -43,7 +43,10 @@ export class SimpleMockSdk extends EventEmitter {
     log: jest.fn(),
     info: jest.fn(),
     warn: jest.fn(),
-    error: jest.fn()
+    error: jest.fn(),
+    config: {
+      appName: 'MockSdkApp'
+    }
   };
   _streamingConnection = {
     config: {
@@ -64,7 +67,8 @@ export class SimpleMockSdk extends EventEmitter {
     },
     _webrtcSessions: {
       refreshIceServers: jest.fn(),
-      getSessionManager: jest.fn()
+      getSessionManager: jest.fn(),
+      proxyNRStat: jest.fn()
     }
   };
   sessionManager = {